Rename ImageJItkService -> SimpleITKService
We can drop "ImageJ" from the name, because of course this is an
ImageJ-related service. It implements the ImageJService interface.

We add "Simple" to the name because this service is for working with
SimpleITK, not just plain ITK, as discussed on GitHub:

https://github.com/imagej/imagej-itk/pull/4#issuecomment-140049410

And we use ITK all upper case, as is always the SciJava convention
for acronyms in class names.